2.6

+++ marquee/TimeClient.cpp
@@ -148,7 +148,7 @@ long TimeClient::getCurrentEpoch() {
 }
 
 long TimeClient::getCurrentEpochWithUtcOffset() {
-  return round(getCurrentEpoch() + 3600 * myUtcOffset + 86400L) % 86400L;
+  return (long)round(getCurrentEpoch() + 3600 * myUtcOffset + 86400L) % 86400L;
 }
